+++ app/system/Bakery/BaseCommand.php
@@ -49,4 +49,19 @@ abstract class BaseCommand extends Command
     {
         $this->ci = $ci;
     }
+
+    /**
+     * Return if the app is in production mode
+     *
+     * @access protected
+     * @return bool
+     */
+    protected function isProduction()
+    {
+        // N.B.: Need to touch the config service first to load dotenv values
+        $config = $this->ci->config;
+        $mode = getenv("UF_MODE") ?: '';
+
+        return ($mode == "production");
+    }
 }